+++ src/private/WRLDPolygon.mm
@@ -79,7 +79,7 @@
              green:&green
               blue:&blue
              alpha:&alpha];
-    return Eegeo::v4(red, green, blue, alpha);
+    return Eegeo::v4(static_cast<float>(red), static_cast<float>(green), static_cast<float>(blue), static_cast<float>(alpha));
 }
 
 - (void)setElevation:(CLLocationDistance)elevation
